+++ index.js
@@ -1,7 +1,7 @@
 var assert = require('assert')
 var qs = require('qs')
 var url = require('url')
-var parse = require('body/json')
+var rawBody = require('raw-body')
 var createRouter = require('wayfarer')
 var isType = require('type-is')
 var httplogger = require('pino-http')
@@ -13,19 +13,33 @@ var error = require('./error')
 /**
 * Create the application. Returns the `app` function that can be passed into `http.createServer`.
 * @name createApp
-* @param {Object} options
+* @param {Object} config
+* @param {Object|Boolean} config.log – Set to `false` to disable logging. Set to an object with options that are passed to [pino](https://npmjs.com/pino) and [pino-http](https://npmjs.com/pino-http)
+* @param {Function} config.notFound – Function that handlers 404 routes. Provides `req`, `res`, and `ctx` arguments just like other appa route handlers. Default: a function that sends a `404` statusCode with the message `Not found`.
+* @returns {Function} `app`
+* @example
+* var appa = require('appa')
+*
+* var app = appa({
+*   log: { level: 'info' }, // or set to `false` to disable all logging
+*   notFound: function (req, res, ctx) {
+*     return app.error(404, 'Not Found').pipe(res)
+*   }
+* })
 */
-module.exports = function createApp (options) {
-  options = options || {}
-  options.log = options.log || { level: 'info' }
+module.exports = function createApp (config) {
+  config = config || {}
 
+  config.log = config.log === false
+    ? { level: 'silent' }
+    : config.log || { level: 'info' }
+
+  var log = createLogger(config.log)
+  var httplog = httplogger(config.log, config.log.stream)
   var router = app.router = createRouter('/404')
 
-  var log = createLogger(options.log)
-  var httplog = httplogger(options.log, options.log.stream)
-
   // provide a 404 fallback
-  on('/404', (options.notFound || notFound))
+  on('/404', (config.notFound || notFound))
   function notFound (req, res) { error('Not found').pipe(res) }
 
   // ignore favicon.ico requests
@@ -37,60 +51,79 @@ module.exports = function createApp (options) {
   * @name app
   * @param {Object} req – the http request object
   * @param {Object} res – the http response object
+  * @example
+  * var http = require('http')
+  * var appa = require('appa')
+  *
+  * var app = appa()
+  * var server = http.createServer(app)
   */
   function app (req, res) {
     httplog(req, res)
-    var parsed = url.parse(req.url)
-    parsed.query = qs.parse(parsed.query)
-    return router(parsed.pathname, req, res, parsed)
+    var ctx = url.parse(req.url)
+    ctx.query = qs.parse(ctx.query)
+    return router(ctx.pathname, req, res, ctx)
   }
 
   /**
   * Route handler
   * @name app.on
   * @param {String} pathname – the route for this handler
+  * @param {Object} options – options for the request handler
+  * @param {Boolean} options.parseJSON – optionally disable JSON parsing of the body. Default: `true`
   * @param {Function} callback – the route handler
+  * @example
+  * app.on('/', function (req, res, ctx) {
+  *   // handle the route
+  * })
   */
-  function on (pathname, callback) {
+  function on (pathname, options, callback) {
+    if (typeof options === 'function') {
+      callback = options
+      options = {}
+    }
+
     assert.equal(typeof pathname, 'string', 'appa: pathname is required and must be a string')
+    assert.equal(typeof options, 'object', 'appa: options must be an object')
     assert.equal(typeof callback, 'function', 'appa: callback function is required')
+
+    options.parseJSON = options.parseJSON === false ? options.parseJSON : true
 
     return router.on(pathname, function (params, req, res, ctx) {
       ctx.params = params
       log.info(ctx)
 
-      function handleParse (err, body) {
-        if (err) return error(400, 'Bad Request, invalid JSON').pipe(res)
-        ctx.body = body
-        callback(req, res, ctx)
-      }
-
       if (req.method === 'POST' || req.method === 'PUT' || req.method === 'PATCH') {
-        if (isType(req, ['json'])) return parse(req, res, handleParse)
-        return callback(req, res, ctx)
+        body(req, options, function (err, result) {
+          if (err) return error(err.status, err.type).pipe(res)
+
+          if (isType(req, ['json']) && options.parseJSON) {
+            ctx.body = parseJSON(res, result)
+          } else {
+            ctx.body = result
+          }
+
+          return callback(req, res, ctx)
+        })
       }
 
-      callback(req, res, ctx)
+      return callback(req, res, ctx)
     })
   }
 
-  /**
-  * Create logs using the pino module: https://npmjs.com/pino
-  * @name app.log
-  */
-  app.log = log
+  function parseJSON (res, result) {
+    try {
+      return JSON.parse(result)
+    } catch (err) {
+      return error(400, 'Invalid JSON').pipe(res)
+    }
+  }
 
-  /**
-  * Parse or stringify a JSON stream using the JSONStream module: https://npmjs.com/JSONStream
-  * @name app.json
-  */
-  app.json = require('JSONStream')
-
-  /**
-  * Compose a stream using the pump module: https://npmjs.com/pump
-  * @name app.pipe
-  */
-  app.pipe = require('pump')
+  function body (req, options, callback) {
+    if (!options.limit) options.limit = '1mb'
+    if (!options.encoding) options.encoding = 'utf8'
+    rawBody(req, options, callback)
+  }
 
   /**
   * Send a JSON object as a response
@@ -120,6 +153,24 @@ module.exports = function createApp (options) {
   * })
   */
   app.error = error
+
+  /**
+  * Create logs using the pino module: https://npmjs.com/pino
+  * @name app.log
+  */
+  app.log = log
+
+  /**
+  * Parse or stringify a JSON stream using the JSONStream module: https://npmjs.com/JSONStream
+  * @name app.json
+  */
+  app.json = require('JSONStream')
+
+  /**
+  * Compose a stream using the pump module: https://npmjs.com/pump
+  * @name app.pipe
+  */
+  app.pipe = require('pump')
 
   app.on = on
   return app