+++ tests/index.js
@@ -11,7 +11,7 @@ function createServer (app) {
 }
 
 test('create a server', function (t) {
-  var app = createApp({ log: { level: 'silent' } })
+  var app = createApp({ log: false })
   var server = createServer(app).listen(0, function () {
     t.ok(app)
     server.close()
@@ -21,7 +21,7 @@ test('create a server', function (t) {
 
 test('create a route', function (t) {
   t.plan(6)
-  var app = createApp({ log: { level: 'silent' } })
+  var app = createApp({ log: false })
 
   app.on('/', function (req, res, context) {
     t.ok(req)